remove animation frames
